requirements.txt
# Web framework
Flask>=2.0.0

# Data handling
pandas>=1.3.0

# Weather API client
openmeteo-requests>=1.0.0
requests-cache>=0.9.0
retry-requests>=2.0.0

# Timezone handling
timezonefinder>=6.0.0
pytz>=2021.1

# Task scheduling
APScheduler>=3.9.1

# Image handling for icon generation
Pillow>=9.0.0

# Optional: For development
pytest>=7.0.0
flake8>=4.0.0
